refactor(scripts): streamline visual regression gates
