+++ test/unit/viewer/ViewerSource.test.ts
@@ -167,6 +167,8 @@ describe("example viewer source", () => {
     const sceneSetupSource = await readFile("examples/viewer/lib/scene-setup.js", "utf8");
     const stateSource = await readFile("examples/viewer/lib/state.js", "utf8");
     const selfShadowGateSource = await readFile("scripts/visual-regression/check-viewer-self-shadow.mjs", "utf8");
+    const selfShadowAnalysisSource = await readFile("scripts/visual-regression/viewer-self-shadow-analysis.mjs", "utf8");
+    const selfShadowGateContracts = `${selfShadowGateSource}\n${selfShadowAnalysisSource}`;
     const styles = await readFile("examples/viewer/styles.css", "utf8");
 
     expect(html).toContain('const threeBuild = baseline');
@@ -324,35 +326,38 @@ describe("example viewer source", () => {
     expect(pipelineSource).toContain("export function setMmdTslDedicatedRawVisibilityDebug(enabled = true)");
     expect(pipelineSource).toContain("mmdTslSelfShadowPass.setReceiverVisibilityDebug(");
     expect(pipelineSource).toContain("export function syncMmdTslDedicatedRawVisibilityDebug()");
-    expect(selfShadowGateSource).toContain('options.backend === "baseline"');
-    expect(selfShadowGateSource).toContain('backend === "webgpu"');
-    expect(selfShadowGateSource).toContain('options.backend = value.toLowerCase()');
-    expect(selfShadowGateSource).toContain("--dedicated-raw-visibility requires --backend webgpu.");
-    expect(selfShadowGateSource).toContain("renderer?.isWebGLRenderer === true");
-    expect(selfShadowGateSource).toContain("renderer?.isWebGPURenderer === true");
-    expect(selfShadowGateSource).toContain("typeof renderer.renderAsync === \"function\"");
-    expect(selfShadowGateSource).toContain("--raw-visibility requires a TSL backend");
-    expect(selfShadowGateSource).toContain("--vmd-lifecycle");
-    expect(selfShadowGateSource).toContain("vmdLifecyclePixelPass(fullFrame");
-    expect(selfShadowGateSource).toContain("vmdInactiveMeanDarkeningMax");
-    expect(selfShadowGateSource).toContain("mode === 0");
-    expect(selfShadowGateSource).toContain("mode === 1 || mode === 2");
-    expect(selfShadowGateSource).toContain("vmdObservation(primary.on.diagnostics)");
-    expect(selfShadowGateSource).toContain('status: "not-applicable"');
-    expect(selfShadowGateSource).toContain("VMD self-shadow mode disabled");
-    expect(selfShadowGateSource).toContain("onModeDistanceAgreement");
-    expect(selfShadowGateSource).toContain("localFullFrameP995DarkeningMin");
-    expect(selfShadowGateSource).toContain("localBackgroundMaxDarkeningMin");
-    expect(selfShadowGateSource).toContain("maxDarkening");
-    expect(selfShadowGateSource).not.toContain("Math.max(...darkening");
-    expect(selfShadowGateSource).toContain("metrics.primary.p995Darkening >= p995Min");
-    expect(selfShadowGateSource).toContain("p95 remains diagnostic-only");
-    expect(selfShadowGateSource).toContain("const inactiveVmdMode = options.vmdLifecycle");
-    expect(selfShadowGateSource).toContain("requireCastShadow = true");
-    expect(selfShadowGateSource).toContain("shadowCameraFar");
-    expect(selfShadowGateSource).toContain("vmdShadowCameraFarPass");
-    expect(selfShadowGateSource).toContain("Math.min(Math.max(on.distance * 100, off.shadowCameraFar), 100)");
-    expect(selfShadowGateSource).toContain("const onShadowCameraFarPass = !enabled");
+    expect(selfShadowGateSource).toContain('from "./viewer-self-shadow-analysis.mjs"');
+    expect(selfShadowAnalysisSource).toContain("export function analyzeReceiverDarkening");
+    expect(selfShadowAnalysisSource).toContain("export function selfShadowDiagnosticsPass");
+    expect(selfShadowGateContracts).toContain('options.backend === "baseline"');
+    expect(selfShadowGateContracts).toContain('backend === "webgpu"');
+    expect(selfShadowGateContracts).toContain('options.backend = value.toLowerCase()');
+    expect(selfShadowGateContracts).toContain("--dedicated-raw-visibility requires --backend webgpu.");
+    expect(selfShadowGateContracts).toContain("renderer?.isWebGLRenderer === true");
+    expect(selfShadowGateContracts).toContain("renderer?.isWebGPURenderer === true");
+    expect(selfShadowGateContracts).toContain("typeof renderer.renderAsync === \"function\"");
+    expect(selfShadowGateContracts).toContain("--raw-visibility requires a TSL backend");
+    expect(selfShadowGateContracts).toContain("--vmd-lifecycle");
+    expect(selfShadowGateContracts).toContain("vmdLifecyclePixelPass(fullFrame");
+    expect(selfShadowGateContracts).toContain("vmdInactiveMeanDarkeningMax");
+    expect(selfShadowGateContracts).toContain("mode === 0");
+    expect(selfShadowGateContracts).toContain("mode === 1 || mode === 2");
+    expect(selfShadowGateContracts).toContain("vmdObservation(primary.on.diagnostics)");
+    expect(selfShadowGateContracts).toContain('status: "not-applicable"');
+    expect(selfShadowGateContracts).toContain("VMD self-shadow mode disabled");
+    expect(selfShadowGateContracts).toContain("onModeDistanceAgreement");
+    expect(selfShadowGateContracts).toContain("localFullFrameP995DarkeningMin");
+    expect(selfShadowGateContracts).toContain("localBackgroundMaxDarkeningMin");
+    expect(selfShadowGateContracts).toContain("maxDarkening");
+    expect(selfShadowGateContracts).not.toContain("Math.max(...darkening");
+    expect(selfShadowGateContracts).toContain("metrics.primary.p995Darkening >= p995Min");
+    expect(selfShadowGateContracts).toContain("p95 remains diagnostic-only");
+    expect(selfShadowGateContracts).toContain("const inactiveVmdMode = options.vmdLifecycle");
+    expect(selfShadowGateContracts).toContain("requireCastShadow = true");
+    expect(selfShadowGateContracts).toContain("shadowCameraFar");
+    expect(selfShadowGateContracts).toContain("vmdShadowCameraFarPass");
+    expect(selfShadowGateContracts).toContain("Math.min(Math.max(on.distance * 100, off.shadowCameraFar), 100)");
+    expect(selfShadowGateContracts).toContain("const onShadowCameraFarPass = !enabled");
     expect(pipelineSource).toContain("state.debugSelfShadowEnabled === true");
     expect(pipelineSource).toContain("mmdTslDedicatedRawVisibilityDebugActive = false;");
     expect(pipelineSource).toContain("if (!root) {");
